+++ Sources/Classes/DataLoading/RxNetworkOperationModel.swift
@@ -81,6 +81,7 @@ open class RxNetworkOperationModel<LoadingStateType: NetworkOperationState>: Net
     func requestResult(from dataSource: DataSourceType) {
         currentRequestDisposable = dataSource
             .resultSingle()
+            .observeOn(MainScheduler.instance)
             .subscribe(onSuccess: { [weak self] result in
                 self?.onGot(result: result, from: dataSource)
             }, onError: { [weak self] error in