+++ TITransitions/Sources/PanelTransition/Animations/PresentAnimation.swift
@@ -1,0 +1,47 @@
+//
+//  Copyright (c) 2020 Touch Instinct
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the Software), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import UIKit
+
+open class PresentAnimation: BaseAnimation {
+    override open func animator(using transitionContext: UIViewControllerContextTransitioning) -> UIViewImplicitlyAnimating {
+        guard let toView = transitionContext.view(forKey: .to),
+            let toController = transitionContext.viewController(forKey: .to) else {
+                return UIViewPropertyAnimator()
+        }
+        
+        let finalFrame = transitionContext.finalFrame(for: toController)
+        
+        toView.frame = finalFrame.offsetBy(dx: .zero, dy: finalFrame.height)
+        
+        let animator = UIViewPropertyAnimator(duration: duration, curve: .easeOut) {
+            toView.frame = finalFrame
+        }
+        
+        animator.addCompletion { _ in
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
+        }
+        
+        return animator
+    }
+}
+